fix(shader): let shader be unique item

+++ Engine/src/Engine/Core/System/Security/Cryptography/Key.h
@@ -6,13 +6,13 @@ namespace Engine
 	class Key
 	{
 	public:
-		Key(uint32_t key)
+		Key(uint64_t key)
 			: m_key(key)
 		{
 		}
 
-		operator uint32_t() const { return m_key; }
+		operator uint64_t() const { return m_key; }
 	private:
-		uint32_t m_key;
+		uint64_t m_key;
 	};
 }
